+++ packages/chains/src/types/chains/index.ts
@@ -1,3 +1,2 @@
 export type { default as Chain } from './Chain';
 export type { default as ChainWithNetworkParameters } from './ChainWithNetworkParameters';
-export type { default as NetworkParametersFromChainParameters } from './NetworkParametersFromChainParameters';