+++ packages/chains/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@kibisis/chains-v1.1.0](https://github.com/kibis-is/toolkit/compare/@kibisis/chains-v1.0.0...@kibisis/chains-v1.1.0) (2025-05-16)
+
+
+### Features
+
+* **@kibisis/chains:** add caip-002 namespace to base chain configurtion ([#3](https://github.com/kibis-is/toolkit/issues/3)) ([956a8b8](https://github.com/kibis-is/toolkit/commit/956a8b8035f6b979e231f4f81f56266ba0d44aa4))
+
 # @kibisis/chains-v1.0.0 (2025-05-15)